Add dictionary module with Dictionary struct to get a vector of strings
This commit adds a module dictionary with the Dictionary struct. The
struct implements a new() function to create a Dictionary struct with an
empty vector and a from_file() function, which reads a file line by line
and push strings in the vector separated by a line. The struct also
stores the maximum word length and alphabet length.