Add bk_tree module and BKTree struct with edit distance and add function
This commit adds the bk_tree module, which contains two structs, Node
and BKTree. The Node struct stores the word along with a vector of
integers storing the indices of words in the BKTree according to the
distance of those words given by the edit function.

The edit distance function is the Damerau Levenshtein distance, which
has been added to the BKTree struct, along with an add() function, which
adds a string into the tre as a Node.