Add does_contain() and get_similar_words() function in BKTree struct
This commit adds two functions: does_contain() and get_similar_words()
in the BKTree struct. The does_contain() function takes in a String
parameter and searches if the word exists in the tree by traversing
through the tree.

The get_similar_words() function takes a String and a tolerance value
and traverses the tree in the specified tolerance level through the root
node and returns a vector of String for each word in the tree that is
equal or less than the tolerance level specified, for the value computed
by the edit function.