Add functions to serialize and deserialize BKTree to and from a file
This commit adds the rkyv crate for serialization. The BKTree and Node
structs in the bk_tree module have been updated with the traits required
to create their archives required for this function. The usize
properties have been changed to fixed sized integer properties to create
the archives.

In addition, two functions from_file() and save_to_file() has been added
for deserializing and serializing the struct respectively.