Add bk tree data structure with functions to add and search words (#1)
This commit adds two modues: bk_tree and dictionary. The dictionary module contains a struct Dictionary which has a vector of String, representing all the words of the dictionary, along with the maximum word length and the alphabet length, which is set during constructing an instance of Dictionary. It has a function to populate the vector of strings from a file separated by a line, with all words switched to lowercase.

The bk_tree module contains a struct BKTree, which has a vector of nodes, as the nodes of the tree. Each node contains a word and an array to the next nodes, according to the edit distance of the words in those nodes. The edit distance is the Damerau-Levenshtein distance, which is a function inside the struct. The BKTree struct also has functions to add a word to the tree, check if a word is present in the tree and to get similar words from the tree with a word to compare as a parameter. The BKTree struct also has functions to serialize the struct to a file or de serialize the struct from a file.

A file dictionary.txt has been added which contains all the words of the English dictionary, separated by a line. Moreover, the Cargo.toml file has been updated to include the readonly and rkyv crates. The rkyv crate has been used for serialization and de-serialization purposes while the readonly crate was used for making the properties of the Dictionary and BKTree struct readonly.