Add bloom filter data structure with implementations for insert, lookup (#2)
This commit adds two modules: bloom_filter and utils.

The bloom_filter module contains a struct BloomFilter and all of its implementations. The insert method calculates several hashes for a string and uses them to insert the word in the BloomFilter. The lookup method is used to search if a word exists in BloomFilter.

The utils module contains handy utility functions for hashing strings, serializing and deserializing data structures.