Add processor module with get_words() and split_words() functions
This commit adds the processor module with two functions: get_words()
and split_words(). The get_words() functions takes a string and then
returns a Vector of String, with each string separated by a whitespace.
The split_words() functions takes a String then returns a tuple of three
strings: the starting punctuations, the middle word, and the ending
punctuations.