Add spell_check module with SpellCheck struct and new() function
This commit adds the spell_check module with the SpellCheck struct. The
struct contains a BKTree and a new() function as implementation. The
new() function currently runs the added get_bk_tree() function to
construct the BKTree.

The get_bk_tree() functions checks if the BKTree is serialized to a
file, if yes then it de-serializes the BKTree from the file. Otherwise,
it creates a Dictionary struct from the dictionary.txt file and then
constructs the BKTree from there, serializes it to a file and returns
it.