Add run() function in SpellCheck struct to run the program
This commit adds the run() function in the SpellCheck struct which runs
a loop to input a paragraph from the user. Then it splits it into a
Vector of String separated by whitespaces. Then each word is split by
removing the starting and ending punctuations.

The word is then passed on to the bloom filter. If the contains()
function of bloom filter is true, then the word is searched in the bk
tree and if it is not present, then the misspelt word is printed out
along with the suggestions. If the contains() function returns false,
then the word isn't searched in the bk tree, since it is already flagged
as a misspelt word. Instead, the word is printed out along with the
suggestions.