Refactor bk_tree module
This commit refactors the bk_tree module by implementing the From trait
for a File for the BKTree struct. Refactoring has been done in all the
functions as well to return the type instead of an Option or Result and
handling errors inside the function. The tests have also been
refactored.