Refactor bk_tree module and optimize BKTree creation from Dictionary
This commit refactors the bk_tree module by replacing the strings in the
nodes with Option<Rc<String>>. In this way, the strings are shared
references instead of clones. A From trait has also been implemented to
construct a BKTree from a Dictionary reference, by cloning each
individual Rc of String instead of the String itself.