Refactor processor and spell_check module
This commit refactors the processor and spell_check module. The
processor module has been refactored to have just two public functions.
One function to split an input string into a vector of tuple of three
String, representing the starting and ending punctuations and the middle
word. Another function to join the words back together.

The spell_check module has been refactored where the BKTree struct and
the BloomFilter struct no longer are under Option and are created in the
constructor itself by passing the file paths. A run() function has been
added which just prints the name of the application for now.