Add processor and spell_check modules and refactor entire code (#3)
This commit adds the processor and spell_check modules. The processor module contains functions to process the input string and the output result. The spell_check modules contains the BKTree and the BloomFilter structs, with its new() function constructing them from the file or from a dictionary.

The processor module contains two public functions: split_input() and join_input(). The first function takes in a String and splits it into a Vector of a tuple of three strings, denoting the starting punctuations, the middle word and the ending punctuations respectively. The second function takes that vector (where the misspelt middle words have been replaced) and returns a string joining them.

The Dictionary, BloomFilter and BKTree structs have been refactored for better error handling, more defined types and in some of the modules, the From trait has been implemented for some types for better construction. The dictionary.txt file have also been refactored to remove duplicate words.